chore: move the cutedsl la_decode to cula.ops
